backend/trane/realtime_dsp/requirements_dsp.txt
Flask>=2.0.0
librosa>=0.10.0
numpy>=1.24.0
soundfile>=0.12.0
celery>=5.3.0
redis>=5.0.0
scipy>=1.11.0
torch>=2.0.0  # For GPU acceleration
webrtcvad>=2.0.10  # For voice activity detection
python-socketio>=5.8.0  # For WebSocket support
psutil>=5.9.0  # For system resource monitoring
py-spy>=0.3.0  # For performance profiling
gputil>=1.4.0  # For GPU monitoring (optional)
nvidia-ml-py3>=7.352.0  # For NVIDIA GPU monitoring (optional)
cachetools>=5.0.0  # For efficient caching
numba>=0.56.0  # For JIT compilation of critical code paths
psycopg2-binary>=2.9.9  # PostgreSQL adapter
django-environ>=0.11.2  # For environment variables
